move tokenbutton into own file
